+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.0.2 (TBD)
+
+* Make `stop` config parameter optional
+* Support for Ollama generate API
+
 ## 0.0.1 (Jan 7, 2025)
 
 * Complete support for llama-cpp and Mistral APIs